backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
google-generativeai
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
